# store: Preallocate output buffer when encoding postings.