--- a/pkg/store/bucket_test.go
+++ b/pkg/store/bucket_test.go
@@ -6,6 +6,7 @@ package store
 import (
 	"bytes"
 	"context"
+	"encoding/binary"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -1921,4 +1922,22 @@ func mustMarshalAny(pb proto.Message) *types.Any {
 		panic(err)
 	}
 	return out
+}
+
+func TestBigEndianPostingsCount(t *testing.T) {
+	const count = 1000
+	raw := make([]byte, count*4)
+
+	for ix := 0; ix < count; ix++ {
+		binary.BigEndian.PutUint32(raw[4*ix:], rand.Uint32())
+	}
+
+	p := newBigEndianPostings(raw)
+	testutil.Equals(t, count, p.count())
+
+	c := 0
+	for p.Next() {
+		c++
+	}
+	testutil.Equals(t, count, c)
 }