--- a/pkg/store/postings_codec.go
+++ b/pkg/store/postings_codec.go
@@ -33,8 +33,9 @@ func isDiffVarintSnappyEncodedPostings(input []byte) bool {
 // diffVarintSnappyEncode encodes postings into diff+varint representation,
 // and applies snappy compression on the result.
 // Returned byte slice starts with codecHeaderSnappy header.
-func diffVarintSnappyEncode(p index.Postings) ([]byte, error) {
-	buf, err := diffVarintEncodeNoHeader(p)
+// Values argument is expected number of postings, used for preallocating buffer.
+func diffVarintSnappyEncode(p index.Postings, values int) ([]byte, error) {
+	buf, err := diffVarintEncodeNoHeader(p, values)
 	if err != nil {
 		return nil, err
 	}
@@ -52,8 +53,15 @@ func diffVarintSnappyEncode(p index.Postings) ([]byte, error) {
 
 // diffVarintEncodeNoHeader encodes postings into diff+varint representation.
 // It doesn't add any header to the output bytes.
-func diffVarintEncodeNoHeader(p index.Postings) ([]byte, error) {
+// Values argument is expected number of postings, used for preallocating buffer.
+func diffVarintEncodeNoHeader(p index.Postings, values int) ([]byte, error) {
 	buf := encoding.Encbuf{}
+
+	// This encoding uses around ~1 bytes per posting, but let's use
+	// conservative 1.25 bytes per posting to avoid extra allocations.
+	if values > 0 {
+		buf.B = make([]byte, 0, 5*values/4)
+	}
 
 	prev := uint64(0)
 	for p.Next() {