@pstibrany 2020-06-26
> 💪
> 
> Motto for the week: PREALLOCATE ;p

Indeed :)